requirements.txt
streamlit
moviepy==1.0.3
google-generativeai
edge-tts